+++ Endless Runner/LivingEntity.m
@@ -62,7 +62,7 @@
 
 - (void)flighttimerFireMethod:(NSTimer *)flighttimer{
     if(self.flying == true){
-        double prevspeed = self.speed;
+        float prevspeed = self.speed;
         if (self.node.physicsBody.velocity.dy < 0 && self.node.position.y < [UIScreen mainScreen].bounds.size.height/2){
           self.speed = -(self.node.physicsBody.velocity.dy/10);
           [self jumpEntity];
